skim the fat from gemma_pytorch
